feat(worldgen): extract caravan stall placement to caravan_stall.js and call from town_gen.js
